refactor(server): share constant-time secret compare
