+++ server/src/middleware/auth.ts
@@ -10,6 +10,7 @@ import { isWorkOSApiKeyFormat } from './api-key-format.js';
 import { verifyWorkOSJWT, looksLikeJWT } from '../auth/workos-jwt.js';
 import { storeRefreshedSession, getRefreshedSession, cleanExpiredRefreshes } from '../db/session-refresh-db.js';
 import { getPool } from '../db/client.js';
+import { constantTimeEqual } from '../utils/constant-time-equal.js';
 
 const logger = createLogger('auth-middleware');
 
@@ -649,27 +650,6 @@ function hasValidAdminApiKey(req: Request): boolean {
   // Don't match WorkOS API keys - those are handled separately
   if (isWorkOSApiKeyFormat(token)) return false;
   return constantTimeEqual(token, ADMIN_API_KEY);
-}
-
-/**
- * Constant-time string compare for ASCII secrets. Intended for comparing
- * attacker-supplied input against a fixed server-side secret (e.g. ADMIN_API_KEY).
- * On length mismatch, runs a same-length dummy compare so total work is
- * O(len(input)), not O(len(secret)) — this leaks the attacker's chosen
- * input length (already public) but never the secret's length.
- *
- * Do NOT reuse this to compare two attacker-controlled inputs against each
- * other — the timing model assumes one side is a known-to-server fixed value.
- * `latin1` encoding makes the ASCII-only assumption on the secret explicit.
- */
-export function constantTimeEqual(a: string, b: string): boolean {
-  const aBuf = Buffer.from(a, 'latin1');
-  const bBuf = Buffer.from(b, 'latin1');
-  if (aBuf.length !== bBuf.length) {
-    timingSafeEqual(aBuf, Buffer.alloc(aBuf.length));
-    return false;
-  }
-  return timingSafeEqual(aBuf, bBuf);
 }
 
 /**
