+++ server/tests/unit/auth-constant-time-equal.test.ts
@@ -1,14 +1,6 @@
-import { describe, it, expect, vi } from 'vitest';
+import { describe, it, expect } from 'vitest';
 
-// auth.ts constructs `new WorkOS(process.env.WORKOS_API_KEY!)` at module
-// load. Stub the module so CI runs against the repo-root vitest config
-// (which has no setup file) don't trip on the missing env var. The
-// pattern matches dev-session-signing.test.ts.
-vi.mock('@workos-inc/node', () => ({
-  WorkOS: class { userManagement = {}; organizations = {}; },
-}));
-
-import { constantTimeEqual } from '../../src/middleware/auth.js';
+import { constantTimeEqual } from '../../src/utils/constant-time-equal.js';
 
 describe('constantTimeEqual', () => {
   it('returns true for identical strings', () => {
